refactor: give the provider roster and two copied helpers one owner each
Ultrareview follow-ups, all three duplication findings.

- `all_cover_hosts` walked a hand-written `[OpenLibrary, GoogleBooks,
  Hardcover]` array, and so did its test — so a fourth provider would compile
  clean and be silently absent from both the CSP `img-src` and the cover-fetch
  SSRF allowlist, surfacing only as a blocked image and a 400. That undercut
  this branch's own claim that the two lists can't drift. `MetadataProvider::ALL`
  is now the single roster, read by `all_cover_hosts`, by `catalog`, and by
  their tests; the per-provider facts stay exhaustive matches, so a new variant
  still has to answer for its hosts and its key.
- `bust_query` existed twice with byte-identical bodies and byte-identical
  tests, one of them commented "Mirrors `cover_editor::bust_query`" — the
  signal it should have been a call. It lives on the page module now, which
  both cover surfaces descend from, and keeps one copy of its tests.
- `EMPTY`, the em-dash placeholder, was declared in three sibling modules in
  one directory. One owner on `metadata_search`.
