+++ frontend/src/pages/metadata_edit.rs
@@ -130,6 +130,20 @@ fn MetadataEditForm(book: EbookMetadata, uuid: String) -> Element {
 // loaded book (which already has any prior overrides merged in); server-side
 // merge then preserves prior overrides on untouched fields.
 
+/// Append a cache-busting `v=` query param to `url`, using `&` when `url`
+/// already carries a query string (mobile's [`crate::media_url`] appends
+/// `?token=…`).
+///
+/// Lives on the page module because both cover surfaces need it — the
+/// sidebar's `CoverEditor` and the edition picker's cover row — and they
+/// bust the same `/api/covers/:uuid` URL for the same reason: the route
+/// caches for a day, so nothing short of a changed query string gets the new
+/// bytes.
+fn bust_query(url: &str, bust: u32) -> String {
+    let sep = if url.contains('?') { '&' } else { '?' };
+    format!("{url}{sep}v={bust}")
+}
+
 /// Edited scalar and list fields collected from the form signals before a save.
 struct EditedFields<'a> {
     title: &'a str,
@@ -218,3 +232,23 @@ fn build_overrides(orig: &EbookMetadata, edited: EditedFields<'_>) -> MetadataOv
 
 #[cfg(test)]
 mod tests;
+
+#[cfg(test)]
+mod bust_query_tests {
+    use super::bust_query;
+
+    #[test]
+    fn bust_query_appends_with_question_mark_when_url_has_no_query_string() {
+        assert_eq!(bust_query("/covers/abc.jpg", 3), "/covers/abc.jpg?v=3");
+    }
+
+    #[test]
+    fn bust_query_appends_with_ampersand_when_url_already_has_a_query_string() {
+        // Mobile's `media_url` appends `?token=…`, so the separator has to be
+        // chosen rather than assumed.
+        assert_eq!(
+            bust_query("/covers/abc.jpg?token=xyz", 3),
+            "/covers/abc.jpg?token=xyz&v=3"
+        );
+    }
+}
